+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### main 
 
+* [Android] Fix rare NPE upon native view disposal.
+
 ### 2.7.0-rc.1
 
 * Update Maps SDK to 11.11.0-rc.1